escalate/requirements_developer.txt
pytest-django
beautifulsoup4